feat(user-ui): update basic info
