+++ packages/user-ui/src/pages/projects/_id/components/basic_info/i18n.rs
@@ -1,0 +1,20 @@
+use bdk::prelude::*;
+
+translate! {
+    BasicInfoTranslate;
+
+    main_title: {
+        ko: "소개글",
+        en: "Introduction"
+    }
+
+    public_opinion_committee_title: {
+        ko: "공론 위원회",
+        en: "Public opinion committee"
+    }
+
+    related_materials_title: {
+        ko: "관련 자료",
+        en: "Related materials"
+    }
+}
